fixes for codeclimate
